--- packages/code-du-travail-data/.eslintrc.js
@@ -1,19 +1,0 @@
-module.exports = {
-  extends: "@socialgouv/eslint-config-typescript",
-  parserOptions: {
-    project: "tsconfig.json",
-    sourceType: "module",
-    tsconfigRootDir: __dirname,
-  },
-  rules: {
-    "@typescript-eslint/naming-convention": "warn",
-    "@typescript-eslint/no-unsafe-return": "warn",
-    "@typescript-eslint/no-floating-promises": "warn",
-    "@typescript-eslint/no-unsafe-argument": "warn",
-    "@typescript-eslint/no-namespace": "warn",
-    "@typescript-eslint/no-explicit-any": "warn",
-    "@typescript-eslint/ban-ts-comment": "warn",
-    "@typescript-eslint/no-unnecessary-condition": "warn",
-    "@typescript-eslint/consistent-type-definitions": "off",
-  },
-};
