--- packages/code-du-travail-data/README.md
@@ -1,14 +1,0 @@
-# code-du-travail-data
-
-L'ensemble des documents et données utilisés par l'api est stocké dans une base elasticsearch et peut être mis à jour en lançant le script d'indexation sans avoir à re-déployer l'api ou le frontend.
-
-## Related
-
--   [fiches-vdd](https://github.com/SocialGouv/fiches-vdd)
--   [legi-data](https://github.com/SocialGouv/legi-data)
--   [kali-data](https://github.com/SocialGouv/kali-data)
--   [contributions-data](https://github.com/SocialGouv/contributions-data)
-
-## Schéma
-
-[![Schema](https://mermaid.ink/svg/eyJjb2RlIjoiZ3JhcGggTFJcbiAgICBcbiAgICBcbiAgICBTaXJldDJJZGNjW0FQSSBTaXJldDJJZGNjXVxuICAgIE5MUFtTZXJ2aW5nIE1MXVxuICAgIEZFLS0-U0lSRU5FW0FQSSBTSVJFTkVdXG5cbiAgICBJTkdFU1tJbmdlc3RlciBlc11cbiAgICBBQ0EtLT5JTkdFU1xuICAgIFNZTi0tPklOR0VTXG4gICAgU1RQLS0-SU5HRVNcbiAgICBzdWdnZXN0ZXIuZGF0YS50eHQtLT5JTkdFU1xuICAgIElOR0VTLS0-RWxhc3RpY1NlYXJjaFxuICAgIE5MUC0tPnxWZWN0b3JpemUgZG9jdW1lbnRzfElOR0VTXG4gICAgU0RSLS0-RkUgIFxuXG4gICAgRkVbRnJvbnRlbmQgQ0RUTl1cbiAgICBGRS0tPlNpcmV0MklkY2NcblxuICAgIFBSRUFWSVMyLS0-RkVcbiAgICBQUkVDQVJJVEUyLS0-RkVcbiAgICBMSUNFTkNJRU1FTlQyLS0-RkUgXG4gICAgXG4gICAgIFxuICAgIEFQSVtBUEkgTm9kZUpTIENEVE5dIFxuICAgIEVsYXN0aWNTZWFyY2gtLT5BUElcbiAgICBOTFAtLT58VmVjdG9yaXplIHNlYXJjaHxBUElcblxuICAgIEZFLS0-QVBJXG4gXG4gICAgQUNBW0FQSSBjZHRuLWFkbWluXVxuXG4gICAgc3ViZ3JhcGggR0lUIENEVE4gXG5cbiAgICAgICAgc3ViZ3JhcGggZ2RvY1xuICAgICAgICAgICAgUFJFQVZJUzJbcHJlYXZpcy5kYXRhLmpzb25dXG4gICAgICAgICAgICBQUkVDQVJJVEUyW3ByZWNhcml0ZS5kYXRhLmpzb25dXG4gICAgICAgICAgICBMSUNFTkNJRU1FTlQyW2xpZW5jaWVtZW50LmRhdGEuanNvbl1cbiAgICAgICAgZW5kXG4gXG4gICAgICAgIHN1YmdyYXBoIG1hbnVlbFxuICAgICAgICAgICAgU0RSW3NlcnZpY2VzLWRlLXJlbnNlaWduZW1lbnQuanNvbl1cbiAgICAgICAgICAgIENPVVJSSUVSU1tjb3VycmllcnMuanNvbl1cbiAgICAgICAgICAgIE9VVElMU1tvdXRpbHMuanNvbl1cbiAgICAgICAgICAgIE9VVElMUzJbZXh0ZXJuYWxzLmpzb25dIFxuICAgICAgICBlbmQgXG4gICAgZW5kIFxuICAgIHN1YmdyYXBoIGNkdG4tZXMgIFxuICAgICAgICBTWU5bc3lub255bWVzLmpzb25dXG4gICAgICAgIFNUUFtzdG9wd29yZHMuanNvbl1cbiAgICBlbmQgXG4gIiwibWVybWFpZCI6eyJ0aGVtZSI6ImRlZmF1bHQifSwidXBkYXRlRWRpdG9yIjpmYWxzZX0)](https://mermaid-js.github.io/mermaid-live-editor/#/edit/eyJjb2RlIjoiZ3JhcGggTFJcbiAgICBcbiAgICBcbiAgICBTaXJldDJJZGNjW0FQSSBTaXJldDJJZGNjXVxuICAgIE5MUFtTZXJ2aW5nIE1MXVxuICAgIEZFLS0-U0lSRU5FW0FQSSBTSVJFTkVdXG5cbiAgICBJTkdFU1tJbmdlc3RlciBlc11cbiAgICBBQ0EtLT5JTkdFU1xuICAgIFNZTi0tPklOR0VTXG4gICAgU1RQLS0-SU5HRVNcbiAgICBzdWdnZXN0ZXIuZGF0YS50eHQtLT5JTkdFU1xuICAgIElOR0VTLS0-RWxhc3RpY1NlYXJjaFxuICAgIE5MUC0tPnxWZWN0b3JpemUgZG9jdW1lbnRzfElOR0VTXG4gICAgU0RSLS0-RkUgIFxuXG4gICAgRkVbRnJvbnRlbmQgQ0RUTl1cbiAgICBGRS0tPlNpcmV0MklkY2NcblxuICAgIFBSRUFWSVMyLS0-RkVcbiAgICBQUkVDQVJJVEUyLS0-RkVcbiAgICBMSUNFTkNJRU1FTlQyLS0-RkUgXG4gICAgXG4gICAgIFxuICAgIEFQSVtBUEkgTm9kZUpTIENEVE5dIFxuICAgIEVsYXN0aWNTZWFyY2gtLT5BUElcbiAgICBOTFAtLT58VmVjdG9yaXplIHNlYXJjaHxBUElcblxuICAgIEZFLS0-QVBJXG4gXG4gICAgQUNBW0FQSSBjZHRuLWFkbWluXVxuXG4gICAgc3ViZ3JhcGggR0lUIENEVE4gXG5cbiAgICAgICAgc3ViZ3JhcGggZ2RvY1xuICAgICAgICAgICAgUFJFQVZJUzJbcHJlYXZpcy5kYXRhLmpzb25dXG4gICAgICAgICAgICBQUkVDQVJJVEUyW3ByZWNhcml0ZS5kYXRhLmpzb25dXG4gICAgICAgICAgICBMSUNFTkNJRU1FTlQyW2xpZW5jaWVtZW50LmRhdGEuanNvbl1cbiAgICAgICAgZW5kXG4gXG4gICAgICAgIHN1YmdyYXBoIG1hbnVlbFxuICAgICAgICAgICAgU0RSW3NlcnZpY2VzLWRlLXJlbnNlaWduZW1lbnQuanNvbl1cbiAgICAgICAgICAgIENPVVJSSUVSU1tjb3VycmllcnMuanNvbl1cbiAgICAgICAgICAgIE9VVElMU1tvdXRpbHMuanNvbl1cbiAgICAgICAgICAgIE9VVElMUzJbZXh0ZXJuYWxzLmpzb25dIFxuICAgICAgICBlbmQgXG4gICAgZW5kIFxuICAgIHN1YmdyYXBoIGNkdG4tZXMgIFxuICAgICAgICBTWU5bc3lub255bWVzLmpzb25dXG4gICAgICAgIFNUUFtzdG9wd29yZHMuanNvbl1cbiAgICBlbmQgXG4gIiwibWVybWFpZCI6eyJ0aGVtZSI6ImRlZmF1bHQifSwidXBkYXRlRWRpdG9yIjpmYWxzZX0)
