--- packages/code-du-travail-data/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@cdt/data",
-  "version": "4.87.0",
-  "description": "Ce dépôt de code permet d'indexer différentes sources de données relatives au Code du travail dans Elasticsearch.",
-  "private": true,
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
-  "scripts": {
-    "precommit": "lint-staged",
-    "build": "tsc",
-    "dev": "nodemon --exec 'npm run build' --watch 'src/**' --ext 'ts,json'",
-    "test:watch": "nodemon --exec 'jest' --watch 'src/**' --ext 'ts,json'",
-    "prepush": "yarn test --bail --changedSince=dev",
-    "lint": "eslint \"./src/**/*.{js,ts}\"",
-    "test": "jest",
-    "clean": "rimraf lib",
-    "type-check": "tsc --noEmit",
-    "type-check:watch": "npm run type-check -- --watch"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/SocialGouv/code-du-travail-numerique.git"
-  },
-  "keywords": [],
-  "bugs": {
-    "url": "https://github.com/SocialGouv/code-du-travail-numerique/issues"
-  },
-  "homepage": "https://github.com/SocialGouv/code-du-travail-numerique/tree/master/packages/code-du-travail-data#readme",
-  "devDependencies": {
-    "@socialgouv/eslint-config-typescript": "1.97.3",
-    "eslint": "7.32.0",
-    "jest": "27.3.1",
-    "lint-staged": "12.1.7",
-    "nodemon": "^2.0.12",
-    "prettier": "2.5.1",
-    "ts-jest": "27.0.7",
-    "typescript": "4.6.4"
-  },
-  "lint-staged": {
-    "*.{ts,js}": [
-      "yarn lint --fix"
-    ]
-  }
-}
