--- packages/code-du-travail-data/src/index.ts
@@ -1,5 +1,0 @@
-export * from "./dismissal-process";
-export * from "./prime-precarite";
-export * from "./simulateurs";
-export * from "./tools";
-export * from "./utils";
