--- packages/code-du-travail-data/src/prime-precarite/index.ts
@@ -1,2 +1,0 @@
-export * from "./PrimePrecarite";
-export * from "./types";
