--- packages/code-du-travail-data/src/simulateurs/index.ts
@@ -1,4 +1,0 @@
-export * from "./HeureRechercheEmploi";
-export * from "./PreavisDemission";
-export * from "./PreavisLicenciement";
-export * from "./types";
