--- packages/code-du-travail-data/src/tools/externals.json
@@ -1,58 +1,0 @@
-[
-  {
-    "title": "Simulateur d’activité partielle",
-    "description": "Estimez simplement le montant de l’allocation d’activité partielle.",
-    "url": "http://www.simulateurap.emploi.gouv.fr/",
-    "icon": "Calculator",
-    "action": "Consulter",
-    "id": "11ec155b-f079-4059-8cda-399ed2ef6026"
-  },
-  {
-    "title": "TéléAccords",
-    "description": "Le service de dépôt des accords collectifs d’entreprise",
-    "url": "https://www.teleaccords.travail-emploi.gouv.fr/PortailTeleprocedures",
-    "icon": "Resignation",
-    "action": "Consulter",
-    "id": "a3a1c93b-379e-4da9-b66b-a89bd913ecb4"
-  },
-  {
-    "title": "TéléRC",
-    "description": "Le service de saisie d'une demande d’homologation de rupture conventionnelle",
-    "url": "https://www.telerc.travail.gouv.fr/RuptureConventionnellePortailPublic/jsp/site/Portal.jsp?page_id=20",
-    "icon": "Contract",
-    "action": "Consulter",
-    "id": "0dc4acd6-4be6-422a-a54f-a11a7bdac9a0"
-  },
-  {
-    "title": "SIPSI",
-    "description": "Détachement de travailleurs en France",
-    "url": "https://www.sipsi.travail.gouv.fr/auth/login",
-    "icon": "Depart",
-    "action": "Consulter",
-    "id": "3608e7fd-24f3-443b-a5b2-e07eb2969dff"
-  },
-  {
-    "title": "Élections professionnelles",
-    "description": "Le site dédié aux élections professionnelles et à la représentativité syndicale",
-    "url": "https://www.elections-professionnelles.travail.gouv.fr/web/guest",
-    "icon": "Document",
-    "action": "Consulter",
-    "id": "81125bf7-a65f-46ec-ad87-2968ed7398ff"
-  },
-  {
-    "title": "Index Egapro",
-    "description": "L’outil de calcul de votre index égalité professionnelle femmes-hommes",
-    "url": "https://index-egapro.travail.gouv.fr",
-    "icon": "Calculator",
-    "action": "Calculer",
-    "id": "84377d4a-7c1b-421b-a9be-efadead39e12"
-  },
-  {
-    "title": "Mon compte formation",
-    "description": "Consultez en ligne vos droits à la formation, cherchez et réservez une formation",
-    "url": "https://www.moncompteformation.gouv.fr",
-    "icon": "Formation",
-    "action": "Consulter",
-    "id": "30b4832d-7584-44de-9b93-2ae5ce9dc57c"
-  }
-]
