--- packages/code-du-travail-data/src/tools/externals.ts
@@ -1,4 +1,0 @@
-import Data from "./externals.json";
-import type { ExternalTool } from "./types";
-
-export const externals: ExternalTool[] = Data;
