--- packages/code-du-travail-data/src/tools/index.ts
@@ -1,3 +1,0 @@
-export * from "./externals";
-export * from "./internals";
-export * from "./types";
