--- packages/code-du-travail-data/src/tools/internals.ts
@@ -1,4 +1,0 @@
-import Data from "./internals.json";
-import type { InternalTool } from "./types";
-
-export const internals: InternalTool[] = Data;
