--- packages/code-du-travail-data/tsconfig.json
@@ -1,28 +1,0 @@
-{
-  "compilerOptions": {
-    /* Visit https://aka.ms/tsconfig.json to read more about this file */
-
-    /* Basic Options */
-    "target": "es5",
-    "module": "commonjs",
-    "lib": [
-      "es2019"
-    ],
-    "outDir": "./lib",
-    "rootDir": "src",
-    "declaration": true,
-    /* Strict Type-Checking Options */
-    "strict": true,
-    "noImplicitAny": true,
-    /* Module Resolution Options */
-    "esModuleInterop": true,
-    /* Advanced Options */
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "downlevelIteration": true
-  },
-  "include": [
-    "src/**/*"
-  ]
-}
