+++ packages/code-du-travail-modeles/src/simulators/dossier-licenciement/index.ts
@@ -1,2 +1,3 @@
 export * from "./questionnaire";
 export * from "./types";
+//TODO :bouger sur frontend
