+++ packages/code-du-travail-modeles/src/simulators/heure-recherche-emploi/HeureRechercheEmploi.ts
@@ -1,4 +1,4 @@
+import type { HeuresRechercheEmploi } from "../types";
 import Data from "./heures-recherche-emploi.data.json";
-import type { HeuresRechercheEmploi } from "./types";
 
 export const heuresRechercheEmploiData: HeuresRechercheEmploi = Data;
