+++ packages/code-du-travail-modeles/src/simulators/index.ts
@@ -1,0 +1,1 @@
+export * from "./types";
