+++ packages/code-du-travail-modeles/src/simulators/preavis-demission/PreavisDemission.ts
@@ -1,4 +1,4 @@
+import type { PreavisDemission } from "../types";
 import Data from "./preavis-demission.data.json";
-import type { PreavisDemission } from "./types";
 
 export const preavisDemissionData: PreavisDemission = Data;
