+++ packages/code-du-travail-modeles/src/simulators/preavis-demission/preavis-demission.data.json
