+++ packages/code-du-travail-modeles/src/simulators/preavis-licenciement/PreavisLicenciement.ts
@@ -1,4 +1,4 @@
+import type { PreavisLicenciement } from "../types";
 import Data from "./preavis-licenciement.data.json";
-import type { PreavisLicenciement } from "./types";
 
 export const preavisLicenciementData: PreavisLicenciement = Data;
