+++ packages/code-du-travail-modeles/src/simulators/prime-precarite/PrimePrecarite.ts
@@ -1,4 +1,4 @@
+import type { PrecariteSituation } from "../types";
 import Data from "./precarite.data.json";
-import type { PrecariteSituation } from "./types";
 
 export const primePrecariteData: PrecariteSituation[] = Data;
