+++ packages/code-du-travail-modeles/src/utils/global.ts
