+++ packages/code-du-travail-modeles/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./constants";
+export * from "./global";
 export * from "./notifications";
 export * from "./references";
